get_cert/requirements.txt
azure-functions
cryptography
